+++ src/model/vector/AxeVector.tpp
@@ -3,11 +3,11 @@
 template<typename T>
 T * AxeVector<T>::getAt(int pos) {
     if(pos >= 0){
-        if(pos >= positif.size()) return nullptr;
+        if(pos >= (int) positif.size()) return nullptr;
         return positif.at(pos);
     }else{
         int posx = pos * (-1) -1;
-        if(posx >= negatif.size()) return nullptr;
+        if(posx >= (int) negatif.size()) return nullptr;
         return negatif.at(posx);
     }
 }
@@ -15,11 +15,11 @@ T * AxeVector<T>::getAt(int pos) {
 template<typename T>
 void AxeVector<T>::addElement(int pos, T *element) {
     if(pos >= 0){
-       if(pos >= positif.size()) positif.resize(pos + 1);
+       if(pos >= (int) positif.size()) positif.resize(pos + 1);
        positif.at(pos) = element;
    }else{
        int posx = (pos*-1)-1;//+du 0
-       if(posx >= negatif.size()) {
+       if(posx >= (int) negatif.size()) {
            negatif.resize(posx+1);//ici +1 de avoir une place en plus
        }
        negatif.at(posx) = element;
